package: rename exception package
