--- src/main/java/com/gaaji/townlife/global/exception/internalServer/InternalServerException.java
@@ -1,4 +1,0 @@
-package com.gaaji.townlife.global.exception.internalServer;
-
-public class InternalServerException extends Exception{
-}
