--- src/main/java/com/gaaji/townlife/global/exception/internalServer/entity/EntityException.java
@@ -1,6 +1,0 @@
-package com.gaaji.townlife.global.exception.internalServer.entity;
-
-import com.gaaji.townlife.global.exception.internalServer.InternalServerException;
-
-public class EntityException extends InternalServerException {
-}
