+++ src/main/java/com/gaaji/townlife/global/exceptions/CategoryNameUniqueConstraintException.java
@@ -1,4 +1,4 @@
-package com.gaaji.townlife.global.exception;
+package com.gaaji.townlife.global.exceptions;
 
 public class CategoryNameUniqueConstraintException extends RuntimeException {
     public CategoryNameUniqueConstraintException(String name) {
