+++ src/main/java/com/gaaji/townlife/global/exceptions/api/AbstractApiException.java
@@ -1,4 +1,4 @@
-package com.gaaji.townlife.global.exception.api;
+package com.gaaji.townlife.global.exceptions.api;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
