+++ src/main/java/com/gaaji/townlife/global/exceptions/api/exception/ResourceAuthorizationException.java
@@ -1,4 +1,7 @@
-package com.gaaji.townlife.global.exception.api;
+package com.gaaji.townlife.global.exceptions.api.exception;
+
+import com.gaaji.townlife.global.exceptions.api.AbstractApiException;
+import com.gaaji.townlife.global.exceptions.api.ErrorCode;
 
 public class ResourceAuthorizationException extends AbstractApiException {
     public ResourceAuthorizationException(ErrorCode errorCode) {
