+++ src/main/java/com/gaaji/townlife/global/exceptions/api/exception/ResourceRemoveException.java
@@ -1,4 +1,7 @@
-package com.gaaji.townlife.global.exception.api;
+package com.gaaji.townlife.global.exceptions.api.exception;
+
+import com.gaaji.townlife.global.exceptions.api.AbstractApiException;
+import com.gaaji.townlife.global.exceptions.api.ErrorCode;
 
 public class ResourceRemoveException extends AbstractApiException {
     public ResourceRemoveException(ErrorCode errorCode) {
