+++ src/main/java/com/gaaji/townlife/global/exceptions/api/exception/ResourceSaveException.java
@@ -1,5 +1,7 @@
-package com.gaaji.townlife.global.exception.api;
+package com.gaaji.townlife.global.exceptions.api.exception;
 
+import com.gaaji.townlife.global.exceptions.api.AbstractApiException;
+import com.gaaji.townlife.global.exceptions.api.ErrorCode;
 import lombok.Getter;
 
 @Getter
