+++ src/main/java/com/gaaji/townlife/global/exceptions/api/exception/RestControllerExceptionHandler.java
@@ -1,5 +1,7 @@
-package com.gaaji.townlife.global.exception.api;
+package com.gaaji.townlife.global.exceptions.api.exception;
 
+import com.gaaji.townlife.global.exceptions.api.AbstractApiException;
+import com.gaaji.townlife.global.exceptions.api.ErrorResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
