+++ src/main/java/com/gaaji/townlife/global/exceptions/internalServer/InternalServerException.java
@@ -1,0 +1,4 @@
+package com.gaaji.townlife.global.exceptions.internalServer;
+
+public class InternalServerException extends Exception {
+}
