+++ src/main/java/com/gaaji/townlife/global/exceptions/internalServer/entity/EntityException.java
@@ -1,0 +1,6 @@
+package com.gaaji.townlife.global.exceptions.internalServer.entity;
+
+import com.gaaji.townlife.global.exceptions.internalServer.InternalServerException;
+
+public class EntityException extends InternalServerException {
+}
