+++ src/main/java/com/gaaji/townlife/service/applicationservice/admin/AdminCategoryModifyServiceImpl.java
@@ -1,7 +1,7 @@
 package com.gaaji.townlife.service.applicationservice.admin;
 
-import com.gaaji.townlife.global.exception.api.ApiErrorCode;
-import com.gaaji.townlife.global.exception.api.ResourceNotFoundException;
+import com.gaaji.townlife.global.exceptions.api.ApiErrorCode;
+import com.gaaji.townlife.global.exceptions.api.exception.ResourceNotFoundException;
 import com.gaaji.townlife.service.controller.admin.dto.AdminCategoryModifyDto;
 import com.gaaji.townlife.service.domain.category.Category;
 import com.gaaji.townlife.service.event.CategoryUpdatedEvent;
