+++ src/main/java/com/gaaji/townlife/service/domain/townlife/TownLife.java
@@ -1,8 +1,8 @@
 package com.gaaji.townlife.service.domain.townlife;
 
-import com.gaaji.townlife.global.exception.api.ApiErrorCode;
-import com.gaaji.townlife.global.exception.api.ResourceAlreadyExistException;
-import com.gaaji.townlife.global.exception.api.ResourceRemoveException;
+import com.gaaji.townlife.global.exceptions.api.ApiErrorCode;
+import com.gaaji.townlife.global.exceptions.api.exception.ResourceAlreadyExistException;
+import com.gaaji.townlife.global.exceptions.api.exception.ResourceRemoveException;
 import com.gaaji.townlife.service.domain.BaseEntity;
 import com.gaaji.townlife.service.domain.category.Category;
 import com.gaaji.townlife.service.domain.comment.ParentComment;
